requirements.txt
transformers
sentence-transformers
PyPDF2
faiss-cpu
numpy
